+++ src/app/apps/utils/text-viewer/page.tsx
@@ -1,5 +1,5 @@
-import MODULE_INFO from '@/components/pages/apps/table-viewer/manifest.json'
-import { TableViewerQueryPage } from '@/components/pages/apps/table-viewer/table-viewer-page'
+import MODULE_INFO from '@/components/pages/apps/utils/text-viewer/manifest.json'
+import { TextViewerQueryPage } from '@/components/pages/apps/utils/text-viewer/text-viewer-page'
 import { makeMetaDataFromModule } from '@/lib/metadata'
 import { Suspense } from 'react'
 
@@ -8,7 +8,7 @@ export const metadata = makeMetaDataFromModule(MODULE_INFO)
 export default function Page() {
   return (
     <Suspense fallback={<div>Loading...</div>}>
-      <TableViewerQueryPage />
+      <TextViewerQueryPage />
     </Suspense>
   )
 }